Commit
This assignment seemed a little too hard considered the amount of time we have spent on some of these concepts and languages.  Mine is incomplete as you will see, I worked on it for 5+ hrs and couldn't really complete any of the problems fully.  Please allow for more resources to help us with these problem sets.